Create Jump to definition addon
