+++ addons/addons.json
@@ -123,6 +123,7 @@
   "hide-stage",
   "one-click-studio-invites",
   "find-bar",
+  "jump-to-def",
 
   "// NEW ADDONS ABOVE THIS ↑↑",
   "// Note: these themes need this exact order to work properly,",
