+++ addons/jump-to-def/addon.json
@@ -1,0 +1,18 @@
+{
+  "name": "Jump to block definition",
+  "description": "Middle click or hold shift while clicking a custom block to jump to its definition.",
+  "credits": [
+    {
+      "name": "griffpatch"
+    }
+  ],
+  "tags": ["editor", "codeEditor"],
+  "userscripts": [
+    {
+      "url": "userscript.js",
+      "matches": ["projects"]
+    }
+  ],
+  "dynamicEnable": true,
+  "dynamicDisable": true
+}
